requirements.txt
# Core Discord & API
discord.py>=2.3.2
fastapi>=0.110.1
uvicorn>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.1

# Database & Cache
sqlalchemy>=2.0.29
asyncpg>=0.29.0
alembic==1.13.3
redis>=5.0.3

# LLM Providers
openai>=1.16.2

# Utilities
python-dotenv>=1.0.1
greenlet>=3.0.3
